style: search input 분리
